tu: Implement partial depth/stencil binding
We missed this when bringing up VK_KHR_dynamic_rendering, but it's
possible to only bind stencil or only bind depth in which case it's
valid to write the unbound aspect and it should be ignored. Fortunately
there's a HW feature to make this possible without fiddling with
depth/stencil state.

Fixes: cb0f414b ("tu: Add support for suspending and resuming renderpasses")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/26154>